+++ lib/stimulus_reflex.rb
@@ -9,6 +9,8 @@ require "action_cable"
 require "nokogiri"
 require "cable_ready"
 require "stimulus_reflex/version"
+require "stimulus_reflex/controller_renderer"
+require "stimulus_reflex/component_renderer"
 require "stimulus_reflex/reflex"
 require "stimulus_reflex/element"
 require "stimulus_reflex/channel"
