+++ lib/stimulus_reflex/channel.rb
@@ -41,7 +41,7 @@ class StimulusReflex::Channel < ActionCable::Channel::Base
       broadcast_message subject: "halted", data: data
     else
       begin
-        render_page_and_broadcast_morph reflex, selectors, data
+        broadcast_morphs reflex.morph(selectors), data
       rescue => render_error
         reflex.rescue_with_handler(render_error)
         message = exception_message_with_backtrace(render_error)
@@ -74,59 +74,6 @@ class StimulusReflex::Channel < ActionCable::Channel::Base
     else
       raise ArgumentError.new("wrong number of arguments (given #{arguments.inspect}, expected #{required_params.inspect}, optional #{optional_params.inspect})")
     end
-  end
-
-  def render_page_and_broadcast_morph(reflex, selectors, data = {})
-
-    morphs = if reflex.component?
-      html = render_component(reflex)
-      morphs = selectors.map { |s| [s, html] }
-    else
-      html = render_page(reflex)
-      collect_morphs selectors, html
-    end
-
-    broadcast_morphs morphs, data
-  end
-
-  def commit_session(request, response)
-    store = request.session.instance_variable_get("@by")
-    store.commit_session request, response
-  rescue => e
-    message = "Failed to commit session! #{exception_message_with_backtrace(e)}"
-    logger.error "\e[31m#{message}\e[0m"
-  end
-
-  def render_page(reflex)
-    controller = reflex.request.controller_class.new
-    controller.instance_variable_set :"@stimulus_reflex", true
-    reflex.instance_variables.each do |name|
-      controller.instance_variable_set name, reflex.instance_variable_get(name)
-    end
-
-    controller.request = reflex.request
-    controller.response = ActionDispatch::Response.new
-    controller.process reflex.url_params[:action]
-    commit_session reflex.request, controller.response
-    controller.response.body
-  end
-
-  def render_component(reflex)
-    controller = reflex.request.controller_class.new
-    controller.instance_variable_set :"@stimulus_reflex", true
-    reflex.instance_variables.each do |name|
-      controller.instance_variable_set name, reflex.instance_variable_get(name)
-    end
-
-    controller.request = reflex.request
-    controller.response = ActionDispatch::Response.new
-    controller.view_context.render reflex.component
-  end
-
-  def collect_morphs selectors, html
-    document = Nokogiri::HTML(html)
-    selectors = selectors.select { |s| document.css(s).present? }
-    selectors.map { |s| [s, document.css(s).inner_html] }
   end
 
   def broadcast_morphs(morphs, data)
